backend/requirements.txt
Flask
Flask-Cors
numpy
psycopg2-binary
python-dotenv